% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pond_code.R
\name{generate_pond_code}
\alias{generate_pond_code}
\title{Generate codes from Pond ID names}
\usage{
generate_pond_code(name)
}
\arguments{
\item{name}{}
}
\description{
Generate codes from Pond ID names
}
