% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_aru_deployments.R
\name{read_aru_deployments}
\alias{read_aru_deployments}
\title{Read ARU deployment data from Google Sheets}
\usage{
read_aru_deployments(google_sheet_url)
}
\arguments{
\item{google_sheet_url}{}
}
\description{
Read ARU deployment data from Google Sheets
}
